+++ cynic-codegen/src/fragment_arguments_derive/mod.rs
@@ -7,9 +7,9 @@ pub fn fragment_arguments_derive(ast: &syn::DeriveInput) -> Result<TokenStream, 
     Ok(quote! {
         impl ::cynic::FragmentArguments for #ident {}
 
-        impl ::cynic::FromArguments<#ident> for () {
-            fn from_arguments(_: &#ident) -> () {
-                ()
+        impl<'a> ::cynic::FromArguments<&'a #ident> for &'a () {
+            fn from_arguments(_: &#ident) -> &() {
+                &()
             }
         }
     })