+++ cynic-querygen/src/lib.rs
@@ -150,7 +150,7 @@ pub fn document_to_fragment_structs(
                 lines.push("    }\n".into());
             }
             PotentialStruct::ArgumentStruct(argument_struct) => {
-                lines.push("    #[derive(cynic::FragmentArguments, Clone, Debug)]".into());
+                lines.push("    #[derive(cynic::FragmentArguments, Debug)]".into());
                 lines.push(format!("    pub struct {} {{", argument_struct.name));
 
                 for field in &argument_struct.fields {
@@ -164,7 +164,7 @@ pub fn document_to_fragment_structs(
                 lines.push("    }\n".into());
             }
             PotentialStruct::InputObject(input_object) => {
-                lines.push("    #[derive(cynic::InputObject, Clone, Debug)]".into());
+                lines.push("    #[derive(cynic::InputObject, Debug)]".into());
                 lines.push(format!(
                     "    #[cynic(graphql_type = \"{}\", rename_all=\"camelCase\")]",
                     input_object.name